+++ pdf/lib/widgets/container.dart
@@ -25,7 +25,7 @@ class BoxBorder {
       this.top = false,
       this.right = false,
       this.bottom = false,
-      this.color = PdfColor.black,
+      this.color = PdfColors.black,
       this.width = 1.0})
       : assert(color != null),
         assert(width != null),