+++ pdf/test/type1_test.dart
@@ -34,15 +34,15 @@ void printText(PdfGraphics canvas, String text, PdfFont font, double top) {
     ..setColor(const PdfColor(0.9, 0.9, 0.9))
     ..fillPath()
     ..drawLine(x + metrics.left - deb, y, x + metrics.right + deb, y)
-    ..setColor(PdfColor.blue)
+    ..setColor(PdfColors.blue)
     ..strokePath()
     ..drawLine(x + metrics.left - deb, y + metrics.ascent,
         x + metrics.right + deb, y + metrics.ascent)
-    ..setColor(PdfColor.green)
+    ..setColor(PdfColors.green)
     ..strokePath()
     ..drawLine(x + metrics.left - deb, y + metrics.descent,
         x + metrics.right + deb, y + metrics.descent)
-    ..setColor(PdfColor.purple)
+    ..setColor(PdfColors.purple)
     ..strokePath()
     ..setColor(const PdfColor(0.3, 0.3, 0.3))
     ..drawString(font, fontSize, text, x, y);