+++ pdf/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add Page.orientation to force landscape or portrait
 * Improve MultiPage Widget
 * Convert GridView to a SpanningWidget
+* Add all Material Colors
 
 # 1.3.3
 * Fix a bug with the RichText Widget